changed Point class to hold doubles

+++ src/main/java/frc/robot/trajectory/Point.java
@@ -1,18 +1,18 @@
 package frc.robot.trajectory;
 
 public class Point {
-    int x, y;
+    double x, y;
 
-    public Point (int x, int y){
+    public Point (double x, double y){
         this.x = x;
         this.y = y;
     }
 
-    public int getX(){
+    public double getX(){
         return x;
     }
      
-    public int getY() {
+    public double getY() {
         return y;
     }
 }
